Open the GitHub page when the version is clicked in the Settings screen
This commit modifies the behavior of the version information in the settings screen. Instead of displaying a toast message, clicking on the version now opens the application's GitHub page in the user's browser. It also adds a fallback mechanism to display a snackbar message if no browser is available. Corresponding Chinese translations have been updated.